+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [JsonSchema] [GH#556](https://github.com/janephp/janephp/pull/556) Support for $defs field
 
 ## [6.3.8] - 2021-05-10
 ### Changed